refactor: enum 클래스 constant 패키지로 이동
