+++ src/main/java/org/programmers/springorder/console/Console.java
@@ -1,14 +1,13 @@
 package org.programmers.springorder.console;
 
-import org.programmers.springorder.constant.ErrorMessage;
-import org.programmers.springorder.constant.Message;
+import org.programmers.springorder.constant.*;
 import org.programmers.springorder.dto.customer.CustomerRequestDto;
 import org.programmers.springorder.dto.customer.CustomerResponseDto;
 import org.programmers.springorder.dto.voucher.VoucherRequestDto;
 import org.programmers.springorder.dto.voucher.VoucherResponseDto;
 import org.programmers.springorder.dto.wallet.WalletRequestDto;
 import org.programmers.springorder.model.voucher.VoucherType;
-import org.programmers.springorder.utils.*;
+import org.programmers.springorder.utils.ExceptionHandler;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
