+++ src/main/java/org/programmers/springorder/controller/voucher/VoucherController.java
@@ -2,10 +2,10 @@ package org.programmers.springorder.controller.voucher;
 
 import org.programmers.springorder.console.Console;
 import org.programmers.springorder.constant.Message;
+import org.programmers.springorder.constant.VoucherMenuType;
 import org.programmers.springorder.dto.voucher.VoucherRequestDto;
 import org.programmers.springorder.service.VoucherService;
 import org.programmers.springorder.utils.ExceptionHandler;
-import org.programmers.springorder.utils.VoucherMenuType;
 import org.springframework.stereotype.Controller;
 
 import java.util.UUID;
