+++ src/test/java/org/programmers/springorder/utils/CustomerMenuTypeTest.java
@@ -3,6 +3,7 @@ package org.programmers.springorder.utils;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
+import org.programmers.springorder.constant.CustomerMenuType;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
