+++ src/test/java/org/programmers/springorder/utils/MenuTypeTest.java
@@ -4,6 +4,7 @@ package org.programmers.springorder.utils;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
+import org.programmers.springorder.constant.MenuType;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
